move slider and buttons back to nwp
